moved logic to its own class
